# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 自炊支援・食費節約アプリ API
  description: |
    一人暮らしまたは2〜4人家族の世帯を対象とした自炊支援・食費節約Webアプリケーションのバックエンド API仕様書。
    
    ## 主要機能
    - ユーザー管理（登録、ログイン、プロフィール管理）
    - レシピ管理（作成、検索、編集、削除）
    - レビュー機能（星評価、コメント、通報）
    - スケジュール管理（料理予定と実績）
    - 買い物リスト管理
    - サボり防止アラート
    - 管理者機能
    
    ## 認証
    - AWS Cognito認証を使用
    - アクセストークンの有効期限: 1時間
    - リフレッシュトークンの有効期限: 90日
    - ユーザーIDはJWTトークンのsubクレームから取得（SecurityContext）
    
    ## 多言語対応
    - 日本語（ja）、韓国語（ko）
    - Accept-Languageヘッダーでエラーメッセージの言語を指定
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: Private
servers:
  - url: http://localhost:8080
    description: ローカル開発環境
  - url: https://api-dev.example.com
    description: 開発環境
  - url: https://api.example.com
    description: 本番環境
...
paths:
...
  /api/recipes/{recipeId}/reviews:
    get:
      tags: [Reviews]
      summary: レビュー一覧取得
      description: レシピのレビュー一覧を取得します。
      operationId: getReviews
      security: []
      parameters:
        - $ref: '#/components/parameters/RecipeIdPathForReview'
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReviewResponse'
    post:
      tags: [Reviews]
      summary: レビュー作成
      description: レシピにレビューを投稿します。ユーザーIDはJWTトークンから取得されます。
      operationId: createReview
      parameters:
        - $ref: '#/components/parameters/RecipeIdPathForReview'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/reviews/{reviewId}:
    put:
      tags: [Reviews]
      summary: レビュー更新
      description: レビューを更新します。ユーザーIDはJWTトークンから取得されます。
      operationId: updateReview
      parameters:
        - $ref: '#/components/parameters/ReviewIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Reviews]
      summary: レビュー削除
      description: レビューを削除します。ユーザーIDはJWTトークンから取得されます。
      operationId: deleteReview
      parameters:
        - $ref: '#/components/parameters/ReviewIdPath'
      responses:
        '204':
          description: 削除成功
        '404':
          $ref: '#/components/responses/NotFound'
  /api/reviews/{reviewId}/report:
    post:
      tags: [Reviews]
      summary: レビュー通報
      description: レビューを通報します。通報カウントが3以上になると自動的に非表示になります。
      operationId: reportReview
      parameters:
        - $ref: '#/components/parameters/ReviewIdPath'
      responses:
        '200':
          description: 通報成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ========================================
  # Schedules API
  # ========================================
...
components:
...
  parameters:
...
    RecipeIdPathForReview:
      name: recipeId
      in: path
      required: true
      description: レシピID
      schema:
        type: string
        format: uuid
    ReviewIdPath:
      name: reviewId
      in: path
      required: true
      description: レビューID
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: バリデーションエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: リソースが見つからない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    CreateReviewRequest:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          maxLength: 300
          example: とても美味しかったです！
    ReviewResponse:
      type: object
      properties:
        reviewId:
          type: string
          format: uuid
        recipeId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        status:
          type: string
          enum: [visible, hidden]
        reportedCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ----------------------------------------
    # Schedule Schemas
    # ----------------------------------------
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: バリデーションエラー
        path:
          type: string
          example: /api/users/register

  # ========================================
  # Security Schemes
  # ========================================
  securitySchemes:
    CognitoAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        AWS Cognito JWTトークン認証。
        ログインエンドポイント（/api/users/login）から取得したアクセストークンを使用します。
        ユーザーIDはJWTトークンのsubクレームから自動的に取得されます（SecurityContext）。
        
        例: `Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`

# ========================================
# Global Security (認証不要のエンドポイントは個別にsecurity: []を指定)
# ========================================
security:
  - CognitoAuth: []